requirements.txt
annotated-types==0.7.0
anyio==4.5.0
arrow==1.3.0
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
fastapi==0.115.0
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
fonttools==4.53.1
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.1
oauthlib==3.2.2
packaging==23.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pushbullet.py==0.12.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
text-unidecode==1.3
ttkbootstrap==1.10.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
